requirement.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
defusedxml==0.7.1
distro==1.6.0
Django==3.2.6
django-celery-beat==2.2.1
django-crispy-forms==1.13.0
django-model-utils==4.2.0
django-simple-email-confirmation==0.70
django-timezone-field==4.2.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
docker==5.0.2
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.2
jsonschema==3.2.0
kombu==5.1.0
Markdown==3.3.4
oauthlib==3.1.1
paramiko==2.7.2
Pillow==8.3.1
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
PyNaCl==1.4.0
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
texttable==1.6.4
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
